Enable MMU

+++ lab6/include/uart.h
@@ -21,7 +21,7 @@
 void uart_init();
 char uart_getc();
 void uart_putc(char c);
-void uart_hex(unsigned int h);
+void uart_hex(unsigned long h);
 void uart_puts(const char *s);
 
 void uart_enable_tx_interrupt();
